Added packetizing of data
We can now send and receive byte arrays longer than the maximum packet
data size.  ACKs are sent in order, as well as a FIN to end the stream
and a FINACK to acknowledge.

NOTE: The packets are sent in order, there is no window size or anything
like that (along with no timeouts and no resending).