Added ability to handle corrupted and lost packets
Instead of using a retransmission packet, I decided to simply ignore
corrupted packets.  In the end, the data is retransmitted properly, the
only problem is that this can take more time than simple retransmission
due to the fact that the sender waits for the timeout if the ACK is lost
or corrupted (or the TRN was corrupted).