Fixed protocol issues in server.c
We can't directly send packets because that doesn't conform to the
protocol. Instead, if there is a file error, we should send either a
speical packet or an empty buffer, and right now I'm defaulting to an
empty buffer.  All communcation should be handled through
{send|receive}Bytes.

Also we shouldn't include *.c files.